requirements.txt
numpy
pygad==2.19.2
dwave-hybrid
dimod
dwave-system